requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
Brotli 
brotlipy==0.7.0
certifi 
cffi==1.17.1
charset-normalizer 
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
crcmod==1.7
cryptography==46.0.3
cycler==0.11.0
dpcpp-cpp-rt==2025.3.1
e2cnn==0.2.3
fonttools==4.42.1
geomloss==0.2.6
idna==3.11
importlib-metadata==6.8.0
importlib-resources==6.0.1
intel-cmplr-lib-rt==2025.3.1
intel-cmplr-lib-ur==2025.3.1
intel-cmplr-lic-rt==2025.3.1
intel-opencl-rt==2025.3.1
intel-openmp==2025.3.1
intel-sycl-rt==2025.3.1
jmespath==0.10.0
kiwisolver==1.4.5
Markdown==3.4.4
markdown-it-py==3.0.0
matplotlib==3.5.0
mdurl==0.1.2
mkl==2025.3.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mmcv-full==1.6.0
mmdet==2.28.2
mmengine==0.8.4
mmrotate==0.3.4
model-index==0.1.11
mpmath==1.3.0
numpy 
onemkl-license==2025.3.0
opencv-python==4.8.0.76
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.24
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==2.0.3
Pillow==9.4.0
platformdirs==3.10.0
prettytable==3.8.0
pycocotools==2.0.7
pycparser==2.23
pycryptodome==3.18.0
Pygments==2.16.1
pyOpenSSL==25.3.0
pyparsing==3.0.9
PySocks 
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.28.2
rich==13.4.2
scipy==1.10.1
segment-anything==1.0
setuptools-scm==7.1.0
shapely==2.0.1
six==1.16.0
sympy==1.12
tabulate==0.9.0
tbb==2022.3.0
tcmlib==1.4.1
termcolor==2.3.0
terminaltables==3.1.10
tomli==2.0.1
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.65.2
typing_extensions 
tzdata==2023.3
umf==1.0.2
urllib3==1.26.20
wcwidth==0.2.6
win-inet-pton==1.1.0
yapf==0.40.1
zipp==3.16.2
